pipeline/requirements.txt
# Sentiment Analysis Dependencies
google-cloud-translate>=3.0.0
pandas>=1.5.0
numpy>=1.21.0
transformers>=4.20.0
torch>=1.12.0
nltk>=3.7
spacy>=3.4.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
requests>=2.28.0
